tools/scraper/requirements.txt
playwright==1.48.0
beautifulsoup4==4.12.3
lxml==5.3.0
python-slugify==8.0.4
python-dotenv==1.0.1
mysql-connector-python==9.0.0
requests==2.32.3
urllib3==2.2.3

